Update: full stack changes across FastAPI, Angular, Spring Boot
- API: refactored routes into modular structure (api/routes/), added error handling (api/errors.py)
- Data layer: added Alembic migrations, refactored database/db.py, updated stores
- Frontend: new components (confirm-dialog, toast-container), role guard, error handler service, patient service, models index, updated auth/config/routes
- Spring Boot: GlobalExceptionHandler, updated security config, JWT filter, AuthController, GatewayController, Medecin model/repo
- Tests: added test suite (conftest, API routes, database, agents, integration, password hashing, registration validation, safety)
- Config: .env.example, environment.prod.ts, updated requirements
- Removed: notebooks/segmentation Unet 3D.ipynb, old api/routes.py

+++ spring-backend/src/main/java/com/agent/repository/MedecinRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MedecinRepository extends JpaRepository<Medecin, Long> {
+public interface MedecinRepository extends JpaRepository<Medecin, Integer> {
     Optional<Medecin> findByUsername(String username);
     Optional<Medecin> findByEmail(String email);
     boolean existsByUsername(String username);
